# api/swagger.yaml
...
swagger: "2.0"
info:
  description: "Hypermedia project by Roccamena, Romanò, Buttolo"
  version: "1.0.0"
  title: "Child4Help"
  termsOfService: "http://swagger.io/terms/"
host: "localhost:8080"
basePath: "/api"
...
schemes:
- "https"
- "http"
paths:
...
  /articles/items:
    get:
      tags:
      - "article"
      summary: "Get all articles items"
      operationId: "getArticlesItemsAll"
      produces:
      - "application/json"
      parameters:
      - name: "page"
        in: "query"
        description: "Articles page number to get"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "Successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/ArticleItem"
      x-swagger-router-controller: "Article"
  /article/{id_article}:
    get:
      tags:
      - "article"
      summary: "Get an article by ID"
      operationId: "getArticleByID"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - name: "id_article"
        in: "path"
        description: "ID of article to return"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "Successful operation"
          schema:
            $ref: "#/definitions/Article"
        "400":
          description: "Invalid ID"
        "404":
          description: "Article not found"
      x-swagger-router-controller: "Article"
    delete:
      tags:
      - "article"
      summary: "Delete an article by ID"
      operationId: "deleteArticle"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - name: "id_article"
        in: "path"
        description: "ID of article to be deleted"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "Successful operation"
        "400":
          description: "Invalid ID"
        "404":
          description: "Article not found"
      x-swagger-router-controller: "Article"
  /article:
    post:
      tags:
      - "article"
      summary: "Add a new article"
      operationId: "addArticle"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - in: "body"
        name: "article"
        description: "Article object to be added"
        required: true
        schema:
          $ref: "#/definitions/Article"
      responses:
        "200":
          description: "Successful operation"
        "400":
          description: "Invalid input"
      x-swagger-router-controller: "Article"
    put:
      tags:
      - "article"
      summary: "Update an article"
      operationId: "updateArticle"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - in: "body"
        name: "article"
        description: "Article object to be updated"
        required: true
        schema:
          $ref: "#/definitions/Article"
      responses:
        "200":
          description: "Successful operation"
        "400":
          description: "Invalid input"
        "404":
          description: "Article not found"
      x-swagger-router-controller: "Article"
  /article/{id_article}/related:
    get:
      tags:
      - "article"
      summary: "Get related articles of a selected article by ID"
      operationId: "getRelatedArticlesByID"
      produces:
      - "application/json"
      parameters:
      - name: "id_article"
        in: "path"
        description: "ID of article to search for related"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "Successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/ArticleItem"
        "400":
          description: "Invalid ID"
        "404":
          description: "Article not found"
      x-swagger-router-controller: "Article"
  /article/{id_article}/service:
    get:
      tags:
      - "article"
      summary: "Get service of a selected article by ID"
      operationId: "getArticleServiceByID"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - name: "id_article"
        in: "path"
        description: "ID of article to search for service"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "Successful operation"
          schema:
            $ref: "#/definitions/ServiceItem"
        "400":
          description: "Invalid ID"
        "404":
          description: "Article not found"
      x-swagger-router-controller: "Article"
  /article/{id_article}/event:
    get:
      tags:
      - "article"
      summary: "Get event of a selected article by ID"
      operationId: "getArticleEventByID"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - name: "id_article"
        in: "path"
        description: "ID of article to search for event"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "Successful operation"
          schema:
            $ref: "#/definitions/EventItem"
        "400":
          description: "Invalid ID"
        "404":
          description: "Article not found"
      x-swagger-router-controller: "Article"
...
  /contact/{id_contact}:
    get:
      tags:
      - "contact"
      summary: "Get a contact by ID"
      operationId: "getContactByID"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - name: "id_contact"
        in: "path"
        description: "ID of contact to return"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "Successful operation"
          schema:
            $ref: "#/definitions/Contact"
        "400":
          description: "Invalid ID"
        "404":
          description: "Contact not found"
      x-swagger-router-controller: "Contact"
    delete:
      tags:
      - "contact"
      summary: "Delete a contact by ID"
      operationId: "deleteContact"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - name: "id_contact"
        in: "path"
        description: "ID of contact to be deleted"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "Successful operation"
        "400":
          description: "Invalid ID"
        "404":
          description: "Contact not found"
      x-swagger-router-controller: "Contact"
  /contact:
    post:
      tags:
      - "contact"
      summary: "Add a new contact"
      operationId: "addContact"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - in: "body"
        name: "contact"
        description: "Contact object to be added"
        required: true
        schema:
          $ref: "#/definitions/Contact"
      responses:
        "200":
          description: "Successful operation"
        "400":
          description: "Invalid input"
      x-swagger-router-controller: "Contact"
    put:
      tags:
      - "contact"
      summary: "Update a contact"
      operationId: "updateContact"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - in: "body"
        name: "contact"
        description: "Contact object to be updated"
        required: true
        schema:
          $ref: "#/definitions/Contact"
      responses:
        "200":
          description: "Successful operation"
        "400":
          description: "Invalid input"
        "404":
          description: "Contact not found"
      x-swagger-router-controller: "Contact"
...
definitions:
...
  ServiceItem:
    type: "object"
    required:
    - "id_service"
    properties:
      id_service:
        type: "integer"
        format: "int64"
      name:
        type: "string"
      presentation:
        type: "string"
      photo_url:
        type: "string"
    xml:
      name: "ServiceItem"
    example:
      presentation: "presentation"
      name: "name"
      photo_url: "photo_url"
      id_service: 0
...
  EventItem:
    type: "object"
    required:
    - "id_event"
    properties:
      id_event:
        type: "integer"
        format: "int64"
      name:
        type: "string"
      date:
        type: "string"
        format: "date"
      location:
        type: "string"
      photo_url:
        type: "string"
    xml:
      name: "EventItem"
    example:
      date: "2000-01-23"
      name: "name"
      location: "location"
      id_event: 0
      photo_url: "photo_url"
...
  Article:
    type: "object"
    required:
    - "id_article"
    properties:
      id_article:
        type: "integer"
        format: "int64"
      id_service:
        type: "integer"
        format: "int64"
      id_event:
        type: "integer"
        format: "int64"
      author:
        type: "string"
      publication_date:
        type: "string"
        format: "date"
      email_address:
        type: "string"
        format: "email"
      title:
        type: "string"
      body:
        type: "string"
      photo_url:
        type: "string"
    xml:
      name: "Article"
    example:
      id_article: 0
      email_address: "email_address"
      author: "author"
      publication_date: "2000-01-23"
      id_event: 1
      photo_url: "photo_url"
      title: "title"
      body: "body"
      id_service: 6
  ArticleItem:
    type: "object"
    required:
    - "id_article"
    properties:
      id_article:
        type: "integer"
        format: "int64"
      author:
        type: "string"
      publication_date:
        type: "string"
        format: "date"
      title:
        type: "string"
      photo_url:
        type: "string"
    xml:
      name: "ArticleItem"
    example:
      id_article: 0
      author: "author"
      publication_date: "2000-01-23"
      photo_url: "photo_url"
      title: "title"
  Contact:
    type: "object"
    required:
    - "id_contact"
    properties:
      id_contact:
        type: "integer"
        format: "int64"
      name:
        type: "string"
      phone:
        type: "string"
      email_address:
        type: "string"
        format: "email"
      location:
        type: "string"
    xml:
      name: "Contact"
    example:
      email_address: "email_address"
      phone: "phone"
      name: "name"
      location: "location"
      id_contact: 0
...